+++ docs/meta/changelog.rst
@@ -1,9 +1,16 @@
 ChangeLog
 =========
 
-Latest Version - ``v0.2.5``
+Latest Version - ``v0.2.6``
 
 ------------------
+
+.. confval:: v0.2.6
+
+- Added two new static methods to ``Stark`` class - :meth:`pystark.client.Stark.list_args` and :meth:`pystark.client.Stark.data`
+- Improved ``Stark.log()`` function. Now pass int values for levels. See :meth:`pystark.client.Stark.log`
+- Added docs for class ``Stark`` - :doc:`Read Here </topics/stark>`
+
 
 .. confval:: v0.2.5